requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
requests>=2.28.0
python-dotenv>=1.0.0
# Optional - only needed for prompt rewriting in Scene Creator
google-genai>=1.0.0
